Relocate proc-macro-coverage-hack to tests and undo workspace setup
