+++ asn1rs-macros/Cargo.toml
@@ -23,8 +23,3 @@ debug-proc-macro = []
 asn1rs-model =  { version = "0.2.0", path = "../asn1rs-model" }
 syn = {version = "1.0.17", features = ["full"] }
 quote = "1.0.3"
-
-[dev-dependencies]
-runtime-macros = "0.3.0"
-old-proc-macro2 = { package = "proc-macro2", version = "0.4" }
-old-syn = { package = "syn", version = "0.15", default-features = false, features = ["parsing", "printing", "full"] }
